+++ mxml.h
@@ -37,27 +37,9 @@ extern "C" {
 
 #  define MXML_TAB		8	// Tabs every N columns
 
-#  define MXML_NO_CALLBACK	0	// Don't use a type callback
-#  define MXML_INTEGER_CALLBACK	mxml_integer_cb
-					// Treat all data as integers
-#  define MXML_OPAQUE_CALLBACK	mxml_opaque_cb
-					// Treat all data as opaque
-#  define MXML_REAL_CALLBACK	mxml_real_cb
-					// Treat all data as real numbers
-#  define MXML_TEXT_CALLBACK	0	// Treat all data as text
-#  define MXML_IGNORE_CALLBACK	mxml_ignore_cb
-					// Ignore all non-element content
-
-#  define MXML_NO_PARENT	0	// No parent for the node
-
 #  define MXML_DESCEND		1	// Descend when finding/walking
 #  define MXML_NO_DESCEND	0	// Don't descend when finding/walking
 #  define MXML_DESCEND_FIRST	-1	// Descend for first find
-
-#  define MXML_WS_BEFORE_OPEN	0	// Callback for before open tag
-#  define MXML_WS_AFTER_OPEN	1	// Callback for after open tag
-#  define MXML_WS_BEFORE_CLOSE	2	// Callback for before close tag
-#  define MXML_WS_AFTER_CLOSE	3	// Callback for after close tag
 
 #  define MXML_ADD_BEFORE	0	// Add node before specified node
 #  define MXML_ADD_AFTER	1	// Add node after specified node
@@ -94,6 +76,14 @@ typedef enum mxml_type_e		// The XML node type.
   MXML_TYPE_CUSTOM			// Custom data
 } mxml_type_t;
 
+typedef enum mxml_ws_e			// Whitespace periods
+{
+  MXML_WS_BEFORE_OPEN,			// Callback for before open tag
+  MXML_WS_AFTER_OPEN,			// Callback for after open tag
+  MXML_WS_BEFORE_CLOSE,			// Callback for before close tag
+  MXML_WS_AFTER_CLOSE,			// Callback for after close tag
+} mxml_ws_t;
+
 typedef void (*mxml_custom_destroy_cb_t)(void *);
 					// Custom data destructor
 
@@ -105,22 +95,28 @@ typedef struct _mxml_node_s mxml_node_t;// An XML node.
 typedef struct _mxml_index_s mxml_index_t;
 					// An XML node index.
 
-typedef bool (*mxml_custom_load_cb_t)(mxml_node_t *, const char *);
+typedef bool (*mxml_custom_load_cb_t)(mxml_node_t *node, const char *s);
 					// Custom data load callback function
 
-typedef char *(*mxml_custom_save_cb_t)(mxml_node_t *);
+typedef char *(*mxml_custom_save_cb_t)(mxml_node_t *node);
 					// Custom data save callback function
 
-typedef int (*mxml_entity_cb_t)(const char *);
+typedef int (*mxml_entity_cb_t)(const char *name);
 					// Entity callback function
 
-typedef mxml_type_t (*mxml_load_cb_t)(mxml_node_t *);
+typedef mxml_type_t (*mxml_load_cb_t)(void *cbdata, mxml_node_t *node);
 					// Load callback function
 
-typedef const char *(*mxml_save_cb_t)(mxml_node_t *, int);
+typedef ssize_t (*mxml_read_cb_t)(void *cbdata, void *buffer, size_t bytes);
+					// Read callback function
+
+typedef const char *(*mxml_save_cb_t)(void *cbdata, mxml_node_t *node, mxml_ws_t when);
 					// Save callback function
 
-typedef bool (*mxml_sax_cb_t)(mxml_node_t *, mxml_sax_event_t, void *);
+typedef ssize_t (*mxml_write_cb_t)(void *cbdata, const void *buffer, size_t bytes);
+					// Write callback function
+
+typedef bool (*mxml_sax_cb_t)(void *cbdata, mxml_node_t *node, mxml_sax_event_t event);
 					// SAX callback function
 
 
@@ -129,7 +125,9 @@ typedef bool (*mxml_sax_cb_t)(mxml_node_t *, mxml_sax_event_t, void *);
 //
 
 extern void		mxmlAdd(mxml_node_t *parent, int where, mxml_node_t *child, mxml_node_t *node);
+
 extern void		mxmlDelete(mxml_node_t *node);
+
 extern void		mxmlElementDeleteAttr(mxml_node_t *node, const char *name);
 extern const char	*mxmlElementGetAttr(mxml_node_t *node, const char *name);
 extern const char       *mxmlElementGetAttrByIndex(mxml_node_t *node, int idx, const char **name);
@@ -137,13 +135,17 @@ extern size_t		mxmlElementGetAttrCount(mxml_node_t *node);
 extern void		mxmlElementSetAttr(mxml_node_t *node, const char *name, const char *value);
 extern void		mxmlElementSetAttrf(mxml_node_t *node, const char *name, const char *format, ...) MXML_FORMAT(3,4);
 extern bool		mxmlEntityAddCallback(mxml_entity_cb_t cb);
-extern const char	*mxmlEntityGetName(int val);
 extern int		mxmlEntityGetValue(const char *name);
 extern void		mxmlEntityRemoveCallback(mxml_entity_cb_t cb);
+
 extern mxml_node_t	*mxmlFindElement(mxml_node_t *node, mxml_node_t *top, const char *element, const char *attr, const char *value, int descend);
 extern mxml_node_t	*mxmlFindPath(mxml_node_t *node, const char *path);
+
 extern const char	*mxmlGetCDATA(mxml_node_t *node);
+extern const char	*mxmlGetComment(mxml_node_t *node);
 extern const void	*mxmlGetCustom(mxml_node_t *node);
+extern const char	*mxmlGetDeclaration(mxml_node_t *node);
+extern const char	*mxmlGetDirective(mxml_node_t *node);
 extern const char	*mxmlGetElement(mxml_node_t *node);
 extern mxml_node_t	*mxmlGetFirstChild(mxml_node_t *node);
 extern long		mxmlGetInteger(mxml_node_t *node);
@@ -157,14 +159,20 @@ extern size_t		mxmlGetRefCount(mxml_node_t *node);
 extern const char	*mxmlGetText(mxml_node_t *node, bool *whitespace);
 extern mxml_type_t	mxmlGetType(mxml_node_t *node);
 extern void		*mxmlGetUserData(mxml_node_t *node);
+
 extern void		mxmlIndexDelete(mxml_index_t *ind);
 extern mxml_node_t	*mxmlIndexEnum(mxml_index_t *ind);
 extern mxml_node_t	*mxmlIndexFind(mxml_index_t *ind, const char *element, const char *value);
-extern int		mxmlIndexGetCount(mxml_index_t *ind);
+extern size_t		mxmlIndexGetCount(mxml_index_t *ind);
 extern mxml_index_t	*mxmlIndexNew(mxml_node_t *node, const char *element, const char *attr);
 extern mxml_node_t	*mxmlIndexReset(mxml_index_t *ind);
-extern mxml_node_t	*mxmlLoadFd(mxml_node_t *top, int fd, mxml_load_cb_t cb); extern mxml_node_t	*mxmlLoadFile(mxml_node_t *top, FILE *fp, mxml_load_cb_t cb);
-extern mxml_node_t	*mxmlLoadString(mxml_node_t *top, const char *s, mxml_load_cb_t cb);
+
+extern mxml_node_t	*mxmlLoadFd(mxml_node_t *top, int fd, mxml_load_cb_t load_cb, void *load_cbdata, mxml_sax_cb_t sax_cb, void *sax_cbdata);
+extern mxml_node_t	*mxmlLoadFile(mxml_node_t *top, FILE *fp, mxml_load_cb_t load_cb, void *load_cbdata, mxml_sax_cb_t sax_cb, void *sax_cbdata);
+extern mxml_node_t	*mxmlLoadFilename(mxml_node_t *top, const char *filename, mxml_load_cb_t load_cb, void *load_cbdata, mxml_sax_cb_t sax_cb, void *sax_cbdata);
+extern mxml_node_t	*mxmlLoadIO(mxml_node_t *top, mxml_read_cb_t read_cb, void *read_cbdata, mxml_load_cb_t load_cb, void *load_cbdata, mxml_sax_cb_t sax_cb, void *sax_cbdata);
+extern mxml_node_t	*mxmlLoadString(mxml_node_t *top, const char *s, mxml_load_cb_t load_cb, void *load_cbdata, mxml_sax_cb_t sax_cb, void *sax_cbdata);
+
 extern mxml_node_t	*mxmlNewCDATA(mxml_node_t *parent, const char *string);
 extern mxml_node_t	*mxmlNewCDATAf(mxml_node_t *parent, const char *format, ...) MXML_FORMAT(2,3);
 extern mxml_node_t	*mxmlNewComment(mxml_node_t *parent, const char *comment);
@@ -182,16 +190,17 @@ extern mxml_node_t	*mxmlNewReal(mxml_node_t *parent, double real);
 extern mxml_node_t	*mxmlNewText(mxml_node_t *parent, bool whitespace, const char *string);
 extern mxml_node_t	*mxmlNewTextf(mxml_node_t *parent, bool whitespace, const char *format, ...) MXML_FORMAT(3,4);
 extern mxml_node_t	*mxmlNewXML(const char *version);
+
 extern int		mxmlRelease(mxml_node_t *node);
 extern void		mxmlRemove(mxml_node_t *node);
 extern int		mxmlRetain(mxml_node_t *node);
-extern char		*mxmlSaveAllocString(mxml_node_t *node, mxml_save_cb_t cb);
-extern bool		mxmlSaveFd(mxml_node_t *node, int fd, mxml_save_cb_t cb);
-extern bool		mxmlSaveFile(mxml_node_t *node, FILE *fp, mxml_save_cb_t cb);
-extern size_t		mxmlSaveString(mxml_node_t *node, char *buffer, size_t bufsize, mxml_save_cb_t cb);
-extern mxml_node_t	*mxmlSAXLoadFd(mxml_node_t *top, int fd, mxml_load_cb_t cb, mxml_sax_cb_t sax, void *sax_data);
-extern mxml_node_t	*mxmlSAXLoadFile(mxml_node_t *top, FILE *fp, mxml_load_cb_t cb, mxml_sax_cb_t sax, void *sax_data);
-extern mxml_node_t	*mxmlSAXLoadString(mxml_node_t *top, const char *s, mxml_load_cb_t cb, mxml_sax_cb_t sax, void *sax_data);
+
+extern char		*mxmlSaveAllocString(mxml_node_t *node, mxml_save_cb_t save_cb, void *save_cbdata);
+extern bool		mxmlSaveFd(mxml_node_t *node, int fd, mxml_save_cb_t save_cb, void *save_cbdata);
+extern bool		mxmlSaveFile(mxml_node_t *node, FILE *fp, mxml_save_cb_t save_cb, void *save_cbdata);
+extern bool		mxmlSaveFilename(mxml_node_t *node, const char *filename, mxml_save_cb_t save_cb, void *save_cbdata);
+extern bool		mxmlSaveIO(mxml_node_t *node, mxml_write_cb_t write_cb, void *write_cbdata, mxml_save_cb_t save_cb, void *save_cbdata);
+extern size_t		mxmlSaveString(mxml_node_t *node, char *buffer, size_t bufsize, mxml_save_cb_t save_cb, void *save_cbdata);
 extern bool		mxmlSetCDATA(mxml_node_t *node, const char *data);
 extern bool		mxmlSetCDATAf(mxml_node_t *node, const char *format, ...) MXML_FORMAT(2,3);
 extern bool		mxmlSetComment(mxml_node_t *node, const char *comment);
@@ -212,19 +221,9 @@ extern bool		mxmlSetText(mxml_node_t *node, bool whitespace, const char *string)
 extern bool		mxmlSetTextf(mxml_node_t *node, bool whitespace, const char *format, ...) MXML_FORMAT(3,4);
 extern bool		mxmlSetUserData(mxml_node_t *node, void *data);
 extern void		mxmlSetWrapMargin(int column);
+
 extern mxml_node_t	*mxmlWalkNext(mxml_node_t *node, mxml_node_t *top, int descend);
 extern mxml_node_t	*mxmlWalkPrev(mxml_node_t *node, mxml_node_t *top, int descend);
-
-
-//
-// Semi-private functions...
-//
-
-extern void		mxml_error(const char *format, ...) MXML_FORMAT(1,2);
-extern mxml_type_t	mxml_ignore_cb(mxml_node_t *node);
-extern mxml_type_t	mxml_integer_cb(mxml_node_t *node);
-extern mxml_type_t	mxml_opaque_cb(mxml_node_t *node);
-extern mxml_type_t	mxml_real_cb(mxml_node_t *node);
 
 
 #  ifdef __cplusplus